+++ prisma/migrations/20220501022023_totp_active_flag/migration.sql
@@ -1,0 +1,2 @@
+-- AlterTable
+ALTER TABLE "Totp" ADD COLUMN     "active" BOOLEAN NOT NULL DEFAULT false;