+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- 2 factor authentication via TOTP (mobile authenticator).
 - Link to Chrome extension ina new page in the settings area.
 
 ### Changed